interfaces: leave addresses when (temporarily) bringing down an interface
This pertains to static setups and VIPs (also static) in cases where we
know the interface will be back later.  A bit of juggling is needed to
properly trigger the required cleanups on dynamic addresses and the new
"implicit" full cleanup ($ifacecfg = true) that appears to be desired
for most callers.

PR: https://forum.opnsense.org/index.php?topic=36807.0